Update usage instructions
The copy of the Sao Paulo GTFS dataset that the usage instructions referenced was deleted because it is now available from https://www.sptrans.com.br/desenvolvedores. However, because the dataset is behind a login, and to prevent the usage example breaking again, I've added a copy to a release on Github that can be used by the example. See https://github.com/transitland/gtfs-archives-not-hosted-elsewhere/commit/49832e51af58bda6a2810dd6c539e1ef073e7300.
Also fixed typo in cargo run example