Only use usize indexes in data model
Replace GTFS data in AppData with usize indexes of the corresponding
row in the RawGtfs table so see if this improves loading or closing
performance of the app.
Doesn't seem to have any noticeable effect. The problem is probably that
Druid is creating every single widget for every item in the data.
Instead, only widgets for the top level items should be created up front
and all other widgets created lazily on demand.