Cache map drawing with capture_image_area
Added local copies of druid and piet in order to add capture_image_area
implementation for Cairo from
https://github.com/linebender/piet/pull/513
It works, except when painting from cache the image might get painted
twice. I think this is because the size of the image returned
by capture_image_area is not the same as the rect passed to it.